reduce rs file size 2/3 tests
