+++ integration-tests/rs/Cargo.toml
@@ -18,13 +18,5 @@ workspaces = "0.2.1"
 pkg-config = "0.3.1"
 
 [[example]]
-name = "increment"
-path = "src/increment.rs"
-
-[[example]]
-name = "decrement"
-path = "src/decrement.rs"
-
-[[example]]
-name = "reset"
-path = "src/reset.rs"
+name = "integration-tests"
+path = "src/tests.rs"
