+++ package.json
@@ -11,7 +11,7 @@
     "test:unit": "npm run build && cd contract && npm i && npm run test",
     "test:integration": "npm run test:integration:ts && npm run test:integration:rs",
     "test:integration:ts": "npm run build && ava --verbose",
-    "test:integration:rs": "npm run build && cd integration-tests/rs && cargo run --example increment && cargo run --example decrement && cargo run --example reset"
+    "test:integration:rs": "npm run build && cd integration-tests/rs && cargo run --example integration-tests"
   },
   "devDependencies": {
     "env-cmd": "^10.1.0",
